# laravel-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/categories": {
            "get": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Récupérer la liste des catégories",
                "operationId": "getCategories",
                "responses": {
                    "200": {
                        "description": "Liste des catégories récupérée avec succès"
                    },
                    "500": {
                        "description": "Erreur interne"
                    }
                }
            },
            "post": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Créer une nouvelle catégorie",
                "operationId": "createCategory",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nom de la catégorie",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Électronique"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description de la catégorie",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Catégorie pour les produits électroniques"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Catégorie créée avec succès"
                    },
                    "400": {
                        "description": "Données invalides"
                    }
                }
            }
        },
        "/api/admin/categories/{categorie_id}": {
            "get": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Récupérer les détails d'une catégorie",
                "operationId": "getCategory",
                "parameters": [
                    {
                        "name": "categorie_id",
                        "in": "path",
                        "description": "ID de la catégorie",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Catégorie récupérée avec succès"
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Mettre à jour une catégorie",
                "operationId": "updateCategory",
                "parameters": [
                    {
                        "name": "categorie_id",
                        "in": "path",
                        "description": "ID de la catégorie",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Informatique"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Catégorie pour les produits informatiques"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Catégorie mise à jour avec succès"
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Supprimer une catégorie",
                "operationId": "deleteCategory",
                "parameters": [
                    {
                        "name": "categorie_id",
                        "in": "path",
                        "description": "ID de la catégorie",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Catégorie supprimée avec succès"
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}